Check metadata for handle
Add option to check the DC and MODS metadata for a handle. If a
handle is found in the metadata and the option is checked, the
handle will not be generated.
This is useful when objects are ingested that already have a handle
but with another prefix than the current handle prefix.
Without this option, these ingested object would get a new handle
and therefore have two handles.