Force derivative generation on handle target change.
When the target of a handle has changed, there is no way to update
the handle with the new target through derivative generation.
This commit checks the current target against the new target when
force derivative generation is true, and when they are not equal,
the target of the handle is updated.