Check metadata for handle
Add option to check the DC and MODS metadata for a handle. If a
handle is found in the metadata and the option is checked, the
handle will not be generated.
This is useful when objects are ingested that already have a handle
but with another prefix than the current handle prefix.
Without this option, these ingested object would get a new handle
and therefore have two handles.
When the handle is retrieved from the metadata, it might have another
prefix than the main prefix. Allow for alternative prefixes and
check if the handle from metadata has one of these alternatives.
If not, then the handle is not one of ours and the handle is
created.